broadcast funcionando, muchas cosas proxy, por ahora solo se implementa para un player

+++ Common/src/Socket/SocketWrapper.cpp
@@ -108,10 +108,10 @@ ssize_t SocketWrapper::operator>>(std::string& s) const {
         return 0;
     }
 
-    // Recibimos el vector
+    // Recibimos el string
     s.reserve(size);
     char byte;
-    for (uint8_t i = 0; i < size; i++) {
+    for (unsigned int i = 0; i < size; i++) {
         last_received += this->operator>>(byte);
         if (last_received) {
             received += last_received;
@@ -136,7 +136,7 @@ ssize_t SocketWrapper::operator<<(const std::vector<uint8_t>& v) const {
 
     // Enviamos el vector
     uint8_t byte;
-    for (uint8_t i = 0; i < size; i++) {
+    for (unsigned int i = 0; i < size; i++) {
         byte = v.at(i);
         last_sent = this->operator<<(byte);
         if (last_sent) {
@@ -164,7 +164,7 @@ ssize_t SocketWrapper::operator>>(std::vector<uint8_t>& v) const {
     // Recibimos el vector
     v.reserve(size);
     uint8_t byte;
-    for (uint8_t i = 0; i < size; i++) {
+    for (unsigned int i = 0; i < size; i++) {
         last_received += this->operator>>(byte);
         if (last_received) {
             received += last_received;
